+++ cg3/data_structures/graphs/includes/iterators/graph_nodeiterator.h
@@ -20,7 +20,7 @@ private:
 
     /* Constructors */
 
-    inline NodeIterator(Graph<T>* graph, typename Graph<T>::Node* node, unsigned int pos) :
+    inline NodeIterator(Graph<T>* graph, typename Graph<T>::Node* node, size_t pos) :
         Graph<T>::GenericNodeIterator(graph, node),
         pos(pos) {}
 
@@ -45,7 +45,7 @@ private:
 
     /* Fields */
 
-    unsigned int pos;
+    size_t pos;
 
 };
